small change, moved code into Node.py file
